first attempt at static counter, but seems complicated so need to rethink it

+++ get_next_line/get_next_line.h
@@ -6,7 +6,7 @@
 /*   By: ameechan <ameechan@student.42.ch>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/22 17:28:46 by ameechan          #+#    #+#             */
-/*   Updated: 2024/02/04 11:51:54 by ameechan         ###   ########.fr       */
+/*   Updated: 2024/02/06 18:03:31 by ameechan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,7 +25,7 @@ typedef struct s_list
 	struct	s_list	*next;
 }					t_list;
 char	*get_next_line(int fd);
-int		found_nl_or_eof(char *buffer);
+unsigned long	found_nl_or_eof(char *buffer);
 t_list	**fill_stash(char *buffer, t_list **line_stash);
 char	*split_nl(char *nl_str, t_list **line_stash);
 size_t	ft_strlen(char *s);
